@@ -1,31 +1,60 @@
-import { AppState } from 'src/app/app.state';
-import { AppWindowRef } from 'src/app/app.window-ref';
-import { AppConfiguration } from 'src/app/app-configuration';
-import { AppService } from 'src/app/app.service';
-import { Component, OnInit, Inject, ViewChild, PLATFORM_ID } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
-import { SolrDocument } from 'src/app/shared/solr-document';
-import { File } from 'src/app/shared/file';
-import { NguCarousel, NguCarouselConfig } from '@ngu/carousel';
-import { LicenseDialogComponent } from '../license-dialog/license-dialog.component';
+import { Component, OnInit, Inject, ViewChild, PLATFORM_ID, signal } from '@angular/core';
+import { MatDialogRef, MAT_DIALOG_DATA, MatDialog, MatDialogModule } from '@angular/material/dialog';
+//import { NguCarousel, NguCarouselConfig } from '@ngu/carousel';
 import { isPlatformBrowser } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { MatCardModule } from '@angular/material/card';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { RouterModule } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
+import { FlexLayoutModule } from 'ngx-flexible-layout';
+import { AppConfiguration } from '../../app-configuration';
+import { AppService } from '../../app.service';
+import { AppState } from '../../app.state';
+import { AppWindowRef } from '../../app.window-ref';
+import { SolrDocument } from '../../shared/solr-document';
+import { File } from '../../shared/file';
+import {
+  NguCarousel, 
+  NguCarouselConfig, 
+  NguCarouselDefDirective,
+  // NguCarouselNextDirective,
+  // NguCarouselPrevDirective,
+  // NguItemComponent
+} from '@ngu/carousel';
+import { LicenseDialogComponent } from '../license-dialog/license-dialog.component';
+import { MatButtonModule } from '@angular/material/button';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+
 
 @Component({
+  imports: [
+    TranslateModule, RouterModule, FlexLayoutModule, FormsModule, MatDialogModule,
+    MatCardModule, MatIconModule,  MatButtonModule, MatFormFieldModule, MatInputModule,
+    MatProgressBarModule, MatTooltipModule, MatListModule, MatSelectModule,
+    NguCarousel,
+    NguCarouselDefDirective
+  ],
   selector: 'app-file-viewer',
   templateUrl: './file-viewer.component.html',
   styleUrls: ['./file-viewer.component.scss']
 })
 export class FileViewerComponent implements OnInit {
 
-  showing = false;
+  showing = signal(false);
   // rolling = false;
   // result: any;
 
-  files: File[] = [];
-  selectedFile: File = null; 
+  files = signal<File[]>([]);
+  selectedFile = signal<File>(null); 
 
   currentPage: number = 1;
-  currentPageDisplayed = 1;
+  currentPageDisplayed = signal(1);
   fileid = 0;
   carouselItems: any[] = [];
 
@@ -57,34 +86,34 @@ export class FileViewerComponent implements OnInit {
 
   ngOnInit(): void {
     this.setData();
-    this.service.logViewer(this.data.ident_cely, this.data.entity).subscribe(() => {});
+    this.service.logViewer(this.data.ident_cely, this.data['entity']).subscribe(() => {});
   }
 
   selectFile(file: File, idx: number) {
     // this.carousel.dataSource = [];
-    this.selectedFile = file;
-    setTimeout(() => {
-
-    //this.carousel.dataSource = this.selectedFile.pages;
+    this.selectedFile.set(file);
+    //setTimeout(() => {
       this.currentPage = 1;
+      this.currentPageDisplayed.set(this.currentPage);
       this.setPage();
       this.fileid = idx + new Date().getTime();
-    }, 10);
+    //}, 10);
   }
 
-  public carouselItemsLoad(j) {
+  public carouselItemsLoad(j: number) {
     this.carouselItems = [];
-    const len = this.selectedFile.pages.length;
+    const len = this.selectedFile().pages.length;
     const max = Math.min(len, this.currentPage + 4 );
       for (let i = 0; i < max; i++) {
         this.carouselItems.push(
-          this.selectedFile.pages[i]
+          i
+          //this.selectedFile().pages[i]
         );
       }
   }
 
   downloadUrl() {
-    return this.config.context + '/api/img/full?id=' + this.selectedFile.id;
+    return this.config.context + '/api/img/full?id=' + this.selectedFile().id;
   }
 
   imgPoint(doc: any, size: string) {
@@ -114,7 +143,7 @@ export class FileViewerComponent implements OnInit {
 
           const link = this.windowRef.nativeWindow.document.createElement('a');
           link.href = this.downloadUrl();
-          link.download = this.selectedFile.nazev;
+          link.download = this.selectedFile().nazev;
           link.click();
           this.service.showInfoDialog(this.service.getTranslation('dialog.desc.download_started'), 2000);
         }
@@ -125,47 +154,60 @@ export class FileViewerComponent implements OnInit {
 
   nextPage() {
     this.currentPage++;
-    this.carousel.moveTo(this.currentPage - 1, false);
-
+    this.currentPageDisplayed.set(this.currentPage);
+    this.carousel.moveTo(this.currentPageDisplayed() - 1, false);
   }
 
   prevPage() {
     if (this.currentPage > 1) {
       this.currentPage--;
-      this.carousel.moveTo(this.currentPage - 1, false);
+      this.currentPageDisplayed.set(this.currentPage);
+      this.carousel.moveTo(this.currentPageDisplayed() - 1, false);
     }
   }
 
   gotoPage() {
     
     this.currentPage = parseInt(this.currentPage+'');
-    if (this.currentPage > 0 || this.currentPage < this.selectedFile.rozsah) {
-      this.carousel.moveTo(this.currentPage - 1, false);
+    this.currentPageDisplayed.set(this.currentPage);
+    if (this.currentPage > 0 || this.currentPage < this.selectedFile().rozsah) {
+      this.carousel.moveTo(this.currentPageDisplayed() - 1, false);
     }
   }
 
   setPage() {
+    const regex = /[0-9]/;
+    const matches = regex.exec(this.currentPage+'');
+    if (!matches) {
+      alert('invalid');
+      return;
+    }
     this.currentPage = parseInt(this.currentPage+'');
+    if (this.currentPage < 1 || this.currentPage > this.selectedFile().rozsah) {
+      alert('invalid');
+      return;
+    }
+    this.currentPageDisplayed.set(this.currentPage);
     this.carouselItemsLoad(this.currentPage);
-    setTimeout(() => {
-      if (this.currentPage > 0 || this.currentPage < this.selectedFile.rozsah) {
-        this.carousel.moveTo(this.currentPage - 1, false);
-      }
-    }, 100);
+    // setTimeout(() => {
+    //   if (this.currentPage > 0 || this.currentPage < this.selectedFile().rozsah) {
+    //     this.carousel.moveTo(this.currentPageDisplayed() - 1, false);
+    //   }
+    // }, 100);
   }
 
   setData() {
-
-
-    this.selectedFile = null;
-    this.files = [];
-    this.showing = false;
+    this.selectedFile.set(null);
+    const files: File[] = [];
+    this.showing.set(false);
 
     setTimeout(() => {
       // this.rolling = false;
       // const fs: any[] = JSON.parse(this.data.soubor);
 
-      this.data.soubor.forEach(f => {
+      // this.rolling = false;
+      // const fs: any[] = JSON.parse(this.data.soubor);
+      this.data['soubor'].forEach((f: any) => {
         const file = new File();
         file.id = f.id;
         file.nazev = f.nazev;
@@ -177,22 +219,23 @@ export class FileViewerComponent implements OnInit {
         file.pages = new Array(file.rozsah);
         file.filepath = f.filepath;
         // file.setSize(true);
-        this.files.push(file);
-        this.files.sort((a, b) => {
+        files.push(file);
+        files.sort((a, b) => {
           return a.nazev.localeCompare(b.nazev);
         });
       });
+      this.files.set(files);
       this.fileid = new Date().getTime();
-      this.selectFile(this.files[0], 0);
+      this.selectFile(this.files()[0], 0);
       // this.selectedFile = this.files[0];
       // this.currentPage = 1;
-      this.showing = true;
+      this.showing.set(true);
     }, 10);
 
   }
 
   mimetype() {
-    const s = this.selectedFile.mimetype;
+    const s = this.selectedFile().mimetype;
     if (s.indexOf('/') > 0) {
       return s.split('/')[1].toUpperCase();
     } else {
@@ -204,7 +247,4 @@ export class FileViewerComponent implements OnInit {
   close() {
     this.dialogRef.close();
   }
-
-
-
 }
