@@ -1,13 +1,114 @@
 import { DatePipe } from '@angular/common';
 import { HttpParams } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit, signal } from '@angular/core';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatTooltipModule } from '@angular/material/tooltip';
 import { Title } from '@angular/platform-browser';
-import { ActivatedRoute, Router, Params } from '@angular/router';
-import { EChartsOption } from 'echarts';
-import { AppService } from 'src/app/app.service';
-import { AppState } from 'src/app/app.state';
+import { ActivatedRoute, Router, Params, RouterModule } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
+import { FlexLayoutModule } from 'ngx-flexible-layout';
+import { AppService } from '../../app.service';
+import { AppState } from '../../app.state';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatIconModule } from '@angular/material/icon';
+import { FormsModule } from '@angular/forms';
+import { MatInputModule } from '@angular/material/input';
+import { MatCardModule } from '@angular/material/card';
+import { MatListModule } from '@angular/material/list';
+import { MatSelectModule } from '@angular/material/select';
+import { MAT_MOMENT_DATE_ADAPTER_OPTIONS, MomentDateAdapter, provideMomentDateAdapter } from '@angular/material-moment-adapter';
+import { MatButtonModule } from '@angular/material/button';
+
+import * as echarts from 'echarts/core';
+import { EChartsOption, ECharts } from 'echarts'; 
+import { NgxEchartsDirective, provideEchartsCore } from 'ngx-echarts';
+import { LineChart } from 'echarts/charts';
+import { TooltipComponent, GridComponent, TitleComponent } from 'echarts/components';
+import { CanvasRenderer } from 'echarts/renderers';
+import { LabelLayout } from "echarts/features";
+import { MAT_DATE_LOCALE, DateAdapter, MAT_DATE_FORMATS } from '@angular/material/core';
+echarts.use([CanvasRenderer, LineChart, TooltipComponent, GridComponent, TitleComponent, LabelLayout]);
+import 'moment/locale/cs';
+
+export class MultiDateFormat {
+  value = '';
+  constructor() { }
+  get display() {
+    switch (this.value) {
+      case 'mm.yyyy':
+        return {
+          dateInput: 'MM.YYYY',
+          monthYearLabel: 'MM YYYY',
+          dateA11yLabel: 'MM.YYYY',
+          monthYearA11yLabel: 'MM YYYY',
+        };
+      case 'yyyy':
+        return {
+          dateInput: 'YYYY',
+          monthYearLabel: 'MM YYYY',
+          dateA11yLabel: 'MM.YYYY',
+          monthYearA11yLabel: 'MM YYYY',
+        };
+      default:
+        return {
+          dateInput: 'DD.MM.YYYY',
+          monthYearLabel: 'MMM YYYY',
+          dateA11yLabel: 'LL',
+          monthYearA11yLabel: 'MMMM YYYY',
+        }
+    }
+
+  }
+  get parse() {
+    switch (this.value) {
+      case 'mm.yyyy':
+        return {
+          dateInput: 'MM.YYYY'
+        };
+      case 'yyyy':
+        return {
+          dateInput: 'YYYY'
+        };
+      default:
+        return {
+          dateInput: 'DD.MM.YYYY'
+        }
+    }
+
+  }
+}
 
 @Component({
+  imports: [
+    TranslateModule,
+    RouterModule,
+    FlexLayoutModule,
+    FormsModule,
+    MatProgressBarModule,
+    MatTooltipModule,
+    MatFormFieldModule,
+    MatDatepickerModule,
+    MatIconModule,
+    MatInputModule,
+    MatCardModule,
+    MatListModule,
+    MatSelectModule,
+    MatButtonModule,
+    NgxEchartsDirective
+],
+  providers: [
+    provideEchartsCore({ echarts }),
+    provideMomentDateAdapter(), 
+    
+    { provide: MAT_DATE_LOCALE, useValue: 'cs-CZ' },
+    // {
+    //   provide: DateAdapter,
+    //   useClass: MomentDateAdapter,
+    //   deps: [MAT_DATE_LOCALE, MAT_MOMENT_DATE_ADAPTER_OPTIONS],
+    // },
+    { provide: MAT_DATE_FORMATS, useClass: MultiDateFormat }
+  ],
   selector: 'app-stats',
   templateUrl: './stats.component.html',
   styleUrls: ['./stats.component.scss']
@@ -32,7 +133,7 @@ export class StatsComponent implements OnInit {
   datumod: Date;
   datumdo: Date;
 
-  loading = false;
+  loading = signal(false);
 
   // insoType: string = 'O';
   // extType: string;
@@ -54,6 +155,7 @@ export class StatsComponent implements OnInit {
 
 
   constructor(
+    // @Inject(MAT_DATE_FORMATS) private dateFormatConfig: MultiDateFormat,
     private datePipe: DatePipe,
     private route: ActivatedRoute,
     private router: Router,
@@ -63,10 +165,16 @@ export class StatsComponent implements OnInit {
 
   ngOnInit(): void {
     this.setTitle();
+
+    // Init this.datumod NOW/YEAR - 1YEAR
+    // this.datumod = new Date(new Date().getFullYear() - 1, 0, 1);
+    this.datumod = new Date(new Date().setFullYear(new Date().getFullYear() - 1))
+
     this.service.currentLang.subscribe(res => {
       this.search(this.route.snapshot.queryParams);
       this.setTitle();
     });
+
     this.subs.push(this.route.queryParams.subscribe(val => {
       this.search(val);
     }));
@@ -109,21 +217,21 @@ export class StatsComponent implements OnInit {
     this.router.navigate([], { queryParams: params, queryParamsHandling: 'merge' });
   }
 
-  removeFilter(field: string, $event) {
+  removeFilter(field: string, event: any) {
     const params: any = {};
     params[field] = undefined;
     params.page = 0;
     this.router.navigate([], { queryParams: params, queryParamsHandling: 'merge' });
-    $event.stopPropagation()
+    event.stopPropagation()
   }
 
   search(params: Params) {
-    this.loading = true;
+    this.loading.set(true);
     this.ident_cely = params['ident_cely'];
     this.type = params['type'];
     this.ip = params['ip'];
     this.user = params['user'];
-    this.interval = params['interval'] ? params['interval'] : 'DAY';
+    this.interval = params['interval'] ? params['interval'] : 'WEEK';
     this.entity = params['entity'];
 
     if (params['date']) {
@@ -148,7 +256,7 @@ export class StatsComponent implements OnInit {
       this.setGraphData(resp.facet_counts.facet_ranges.indextime.counts);
 
       this.totalIds = resp.stats.stats_fields.ident_cely.countDistinct;
-      this.loading = false;
+      this.loading.set(false);
     });
   }
 
@@ -201,7 +309,7 @@ export class StatsComponent implements OnInit {
     this.chartOptions.series = this.series;
     this.chartOptions.legend = { data: this.legend, bottom: 0 };
 
-    let title = this.service.getTranslation('stats.graphTitle.' + (this.interval ? this.interval : 'DAY'));
+    let title = this.service.getTranslation('stats.graphTitle.' + (this.interval ? this.interval : 'WEEK'));
     this.chartOptions.title = {
       text: title,
       left: 'center'
@@ -210,3 +318,4 @@ export class StatsComponent implements OnInit {
   }
 
 }
+
